requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.2.35
cryptography==39.0.1
defusedxml==0.7.1
dj-rest-auth==3.0.0
Django==3.2.15
django-allauth==0.52.0
django-celery-beat==2.5.0
django-environ==0.9.0
django-filter==23.1
django-phonenumber-field==7.0.2
django-rest-swagger==2.2.0
django-timezone-field==5.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-extra-fields==3.4.1
drf-psq==1.1.0
drf-spectacular==0.26.0
exceptiongroup==1.1.1
Faker==18.3.2
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
kombu==5.2.4
MarkupSafe==2.1.2
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.0
phonenumberslite==8.13.6
Pillow==9.4.0
pluggy==1.0.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pyrsistent==0.19.3
pytest==7.2.2
python-crontab==2.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
redis==4.5.3
requests==2.28.2
requests-oauthlib==1.3.1
simplejson==3.18.3
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2023.1
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
